@@ -28,9 +28,9 @@
     },
     "minimum-stability": "stable",
     "require": {
-        "php": ">=5.4"
+        "php": "^7.0"
     },
     "require-dev": {
-        "phpunit/phpunit":"~4.3"
+        "phpunit/phpunit":"~6"
     }
 }
