@@ -2,13 +2,13 @@
 /**
  * @group validation
  */
-class NBD_Validation_Rules_AlphaRuleTest extends PHPUnit_Framework_TestCase {
+class NBD_Validation_Rules_AlphaRuleTest extends \PHPUnit\Framework\TestCase {
 
   protected $_class = 'Behance\NBD\Validation\Rules\AlphaRule';
 
   /**
    * @test
-   * @dataProvider testDataProvider
+   * @dataProvider isValidDataProvider
    */
   public function isValid( $data, $expected ) {
 
@@ -23,7 +23,7 @@ class NBD_Validation_Rules_AlphaRuleTest extends PHPUnit_Framework_TestCase {
   /**
    * @return array
    */
-  public function testDataProvider() {
+  public function isValidDataProvider() {
 
     return [
         [ 'abc', true ],
@@ -45,6 +45,6 @@ class NBD_Validation_Rules_AlphaRuleTest extends PHPUnit_Framework_TestCase {
         [ ( function() {} ), false ],
     ];
 
-  } // testDataProvider
+  } // isValidDataProvider
 
 } // NBD_Validation_Rules_AlphaRuleTest
