@@ -2,13 +2,13 @@
 /**
  * @group validation
  */
-class NBD_Validation_Rules_ArrayRuleTest extends PHPUnit_Framework_TestCase {
+class NBD_Validation_Rules_ArrayRuleTest extends \PHPUnit\Framework\TestCase {
 
   protected $_class = 'Behance\NBD\Validation\Rules\ArrayRule';
 
   /**
    * @test
-   * @dataProvider testDataProvider
+   * @dataProvider isValidDataProvider
    */
   public function isValid( $data, $expected ) {
 
@@ -23,7 +23,7 @@ class NBD_Validation_Rules_ArrayRuleTest extends PHPUnit_Framework_TestCase {
   /**
    * @return array
    */
-  public function testDataProvider() {
+  public function isValidDataProvider() {
 
     return [
         [ 'abc', false ],
@@ -48,6 +48,6 @@ class NBD_Validation_Rules_ArrayRuleTest extends PHPUnit_Framework_TestCase {
         [ [ 0 => 'abc', 1 => 123, 'def' => true ], true ],
     ];
 
-  } // testDataProvider
+  } // isValidDataProvider
 
 } // NBD_Validation_Rules_ArrayRuleTest
