@@ -5,11 +5,11 @@ namespace Behance\NBD\Validation\Rules;
 /**
  * @group validation
  */
-class HexColorRuleTest extends \PHPUnit_Framework_TestCase {
+class HexColorRuleTest extends \PHPUnit\Framework\TestCase {
 
   /**
    * @test
-   * @dataProvider testDataProvider
+   * @dataProvider isValidDataProvider
    */
   public function isValid( $data, $expected ) {
 
@@ -23,7 +23,7 @@ class HexColorRuleTest extends \PHPUnit_Framework_TestCase {
   /**
    * @return array
    */
-  public function testDataProvider() {
+  public function isValidDataProvider() {
 
     return [
         [ 'abc', true ],
@@ -58,6 +58,6 @@ class HexColorRuleTest extends \PHPUnit_Framework_TestCase {
         [ ( function() {} ), false ],
     ];
 
-  } // testDataProvider
+  } // isValidDataProvider
 
 } // HexColorRuleTest
