@@ -2,13 +2,13 @@
 /**
  * @group validation
  */
-class NBD_Validation_Rules_MaxLengthRuleTest extends PHPUnit_Framework_TestCase {
+class NBD_Validation_Rules_MaxLengthRuleTest extends \PHPUnit\Framework\TestCase {
 
   protected $_class = 'Behance\NBD\Validation\Rules\MaxLengthRule';
 
   /**
    * @test
-   * @dataProvider testDataProvider
+   * @dataProvider isValidDataProvider
    */
   public function isValid( $data, $max, $expected ) {
 
@@ -76,7 +76,7 @@ class NBD_Validation_Rules_MaxLengthRuleTest extends PHPUnit_Framework_TestCase 
   /**
    * @return array
    */
-  public function testDataProvider() {
+  public function isValidDataProvider() {
 
     return [
         [ 'abc', 4, true ],
@@ -98,6 +98,6 @@ class NBD_Validation_Rules_MaxLengthRuleTest extends PHPUnit_Framework_TestCase 
         [ ( function() {} ), 4, false ],
     ];
 
-  } // testDataProvider
+  } // isValidDataProvider
 
 } // NBD_Validation_Rules_MaxLengthRuleTest
