@@ -2,13 +2,13 @@
 /**
  * @group validation
  */
-class NBD_Validation_Rules_NotEmptyRuleTest extends PHPUnit_Framework_TestCase {
+class NBD_Validation_Rules_NotEmptyRuleTest extends \PHPUnit\Framework\TestCase {
 
   protected $_class = 'Behance\NBD\Validation\Rules\NotEmptyRule';
 
   /**
    * @test
-   * @dataProvider testDataProvider
+   * @dataProvider isValidDataProvider
    */
   public function isValid( $data, $expected ) {
 
@@ -23,7 +23,7 @@ class NBD_Validation_Rules_NotEmptyRuleTest extends PHPUnit_Framework_TestCase {
   /**
    * @return array
    */
-  public function testDataProvider() {
+  public function isValidDataProvider() {
 
     return [
         [ 'abc', true ],
@@ -41,6 +41,6 @@ class NBD_Validation_Rules_NotEmptyRuleTest extends PHPUnit_Framework_TestCase {
         [ null, false ],
     ];
 
-  } // testDataProvider
+  } // isValidDataProvider
 
 } // NBD_Validation_Rules_NotEmptyRuleTest
