@@ -3,13 +3,13 @@
 /**
  * @group validation
  */
-class NBD_Validation_Rules_StringRuleTest extends PHPUnit_Framework_TestCase {
+class NBD_Validation_Rules_StringRuleTest extends \PHPUnit\Framework\TestCase {
 
   protected $_class = '\Behance\NBD\Validation\Rules\StringRule';
 
   /**
    * @test
-   * @dataProvider testDataProvider
+   * @dataProvider isValidDataProvider
    */
   public function isValid( $data, $expected ) {
 
@@ -24,7 +24,7 @@ class NBD_Validation_Rules_StringRuleTest extends PHPUnit_Framework_TestCase {
   /**
    * @return array
    */
-  public function testDataProvider() {
+  public function isValidDataProvider() {
 
     return [
         [ 'abc', true ],
@@ -49,6 +49,6 @@ class NBD_Validation_Rules_StringRuleTest extends PHPUnit_Framework_TestCase {
         [ [ 0 => 'abc', 1 => 123, 'def' => true ], false ],
     ];
 
-  } // testDataProvider
+  } // isValidDataProvider
 
 } // NBD_Validation_Rules_StringRuleTest
